Reorganized the project
Moved the example input files to an examples/ folder sEH_TPPU/

Took the __main__ part out of WExplore and put it into the example as
well. Am working on getting the right functionality out of the example
and into the module.

moved my original prototype examples into a prototypes/ dir.

Started a real development dir wepy/ under the main one.

Made a stub for the setup.py, not sure if it will work with openmm
dependency though.

changed a few other syntaxy things around too that will probably break
things but will be fixed soon.