Working on skeleton, added new classes and stubs
Lots changed to the skeleton of the project.
I made new modules for:
- runner
  - openmm runner
- resampling : the cloning and merging classes
  - I started this but didn't get a working example other than the
  trivial case of no merges and clones
- walker
   - with methods for cloning and merging

The stub example I made is intended to just put all the stubs together
into something that runs.

I made the openmm thing it's own module because I felt wrong
hardcoding all of that in to the sim manager and it really is quite
easy just to put it in another class and have it's run segment method
called from the sim manager.

THis trivial change will make the whole thing modular to custom
engines for whatever really as long as it is callable in python!