Only 1 class for resampling, Resampler
The resampler handles everything from distance metric to decisions and
actually cloning and merging.

Before I was having to pass things like decision vectors between
function calls in the Manager and it seemed unnecessary and limiting
and would force us into a design decision. So there will just be one
class that has a method ~resample~ that takes in a collection of
walkers and returns a new list of walkers and a record of the changes
that were made.

Very simple. That unites a lot of classes just in terms of stubs, and
makes writing of tightly coupled resamplers easier. While maybe that
isn't the best design I think it is better to allow for this in a
scientific setting were people just want to prototype something
without implementing 3 or 4 classes!

However because this is OO you can make your resampler class as
complex or complicated as you could ever dream of. Including other
modules etc.

NAZANIN: I know I told you to write things in the Decision module but
now things will go into the Resampler module.

We can in the future work on helpers for doing specific patterns of
cloning and merging and distance metrics like we were doing before but
for now just copy-paste the code you have and make it return walkers
and a resampling record and we will have a testable prototype!