new OpenMM example, OpenMMWalker, count walker class
Updating the example for openmm to work. Starting just with the
builtin map function.

I made an OpenMMWalker class that uses a context state as the state
and added properties to access its stuff more pythonic and convenient.

There was an issue with mapping the right number of segment_lengths to
the mapper because the walkers to be mapped were in a generator and so
Icould not know the length and I didn't want to keep them in
memory. So I made a wrapper class for the iterator that unpacks them
into memory counts them and makes an index then repackages them back
into a generator which it returns so they are taken out of memory,
maybe... maybe that list is still around IDK? Anyways that is a point
of improvement in the sim_manager implementatino for performance we
can tweak in the future.