started a random resampler for testing, test script, walker.squash
I am trying to get a prototype for resampling records and output from
a resampler and I wanted to have a resampler that does something
interesting and on the surface similar to what would be done in say
WExplore so I made one that just does a random number of clones and
then squashes some other random walker which merges with another
random walker.

In this way it is possible for a walker to have multiple clones occur
to it. So I can analyze this "stack" of actions and then think of the
best way to store this data.

I have made a test script which I havent used yet.

To accomplish this kind of workflow I had to implement an explicit
squash function for a walker as the merge function does stuff with
picking one to merge to. I could incorporate this in another resampler.