Moved clone merge example stuff to its own module
All the stuff related specifically to cloning and merging was moved to
the clone_merge.py module so it is all together and uses the rest of
the library in a canonical way.

Updated the NoResampler to conform to the current format of having
multiple stages to resampling in one resample call.

Made the base Decision have true and false as decisions for testing
etc.

Added Instruction records for each individual enumerated decision that
is specific to a kind of decision making process.