RandomCloneMerge fix, Instruction records
I made a mistake where I was not setting the walkers at the end of a
resample stage to the newly resampled walkers so that they were
getting reset to the initial walkers each stage.

Also implemented the use of the ResamplingRecord and the
DecisionInstructionRecords in the resampler. This just makes it
explicit and semantic as to what the "value" is in a ResamplingRecord,
which is a decision and the instruction with what to do with that.