OpenMMWalker API, resampling records, hdf5 module started
made additions to the OpenMMWalker API by including methods to get
numpy arrays out easily for simtk.units.Quantities which come from
OpenMM states.

Tried to make mdtraj trajectories easily but there is a discrepency
with the box vectors given from OpenMM and mdtraj requiring box
lengths and box angles, lame. So I decided just to roll completely my
own.

Started a todo file.

Made a dataframe and csv out of the resampling record example so there
is an example for that and it can be used for other testing purposes.

Started a new submodule for hdf5 storage format in wepy itself. I will
use the old wepy_hdf5 as a guide but I have found that the underlying
structure is different than before and the old prototype is a bit
messy. The old prototype works for the perl implementation of WExplore
so it can still be used for that.