writing code in TrajHDF5 for adding attributes in different methods
I created a dataset for which I can test things and I was getting lots
of confusin type errors so I am implementing a system that lets you
check for errors in the setter and just validate them in the init
phase.

I broke the init function to different functions and am working on the
write one.

I broke down the attributes into:
- topology; mandatory
- data
  - positions; mandatory
  - time; optional, and the rest
  - etc...
- units

and used a pattern where I put them in a dict and use that dict in a
loop to use the __dict__ and __setattr__ functions to set them and add
them. This gives us the flexibility to easily add new ones in the
future.

I also made the distinction of simple and compound variables.

I am only dealing with simple ones for now but the compound ones are:
- forces
- parameters
- observables

Which will have keys to access specific datasets within them, and
there will be a mirrored units attribute for that as well.