starting to make consistent and new I/O modes
Finished refactoring to match the h5py conventions and added two new
modes as stubs to meet our needs this is the 'c' and 'c-' modes which
allow addition of new datasets and/or appending to existing datasets.

w, w-, and x all just create new files and the a option just does
read/write and creates a new file if needed. it is not append as I was
calling it.