working on WepyHDF5, and Reporters, WepyHDF5Reporter
I got most of the WepyHDF5 core features for runs and trajectories
implemented.

I then started making a reporter module which you give to the
simulation manager and implements an init that is separate from the
constructore __init__. __init__ is used for adding parameters of how
to do things. The init does things to prepare once the simulation has
actually started, such as opening files.

They also have the report function which is called every cycle.

This made me realize I have to handle the weights, which is different
than what is put into the WepyHDF5 traj leaf group but not in the
TrajHDF5. So I added that as well as implementing an append to a
trajectory.

Also to make it easy to dump the state values as a dict I implemented
the __getitem__ and keys for the OpenMM walker so you can **walker
unpack it. This is only for the state variables not the weight.