Working version of HDF5 reporter
The major achievement of this commit is the first working version of
the WepyHDF5Reporter class.

This first required making a new submodule for reporters that includes
some abstract base classes, that really aren't that much use at the
time.

Within the submodule there is a module for HDF5 reporters with one
class for the WepyHDF5 file type.

Right now it implements saving runs and trajectory information.

In future installments I will add the resampling records and the
boundary conditions etc.

This required routing some deficiencies in a number of other classes,
notably the OpenMM walker interface. Which is something that should in
the future be unified and tried to make stable as this could be a
point of trouble.

I also stumbled upon the requirement for cloning and merging to have
at least 3 walkers, so I made checks for that in the clone_merge
module.

The introduction of the reporter needed to be accomodated in the
sim_manager which identified 3 major functions that needed to be
implemented by a reporter:
- init() :: an argumentless post constructor command, used to open files etc.
- report(data) :: the function that actually reports the data somehow
- cleanup() :: a post-simulation tear function with no arguments, used
to close files etc.

This model allows for the reporter constructor to set object variables
like the file path and modes to open it with, as well as potentially
other parameters, while allowing the file to be opened by the
sim_manager.

A few changes were also needed in the WepyHDF5 class as well.