adding resampling to HDF5
Accomodating the resampling takes a bit of effort due to the variable
length of the instructions in the resampling records.

So I have implemented the initialization of the resampling group and
decision instruction datasets in WepyHDF5 for only the fixed width
instructions, not the variable width.

This was done after I set up the surrounding API calls in the
WepyHDF5Reporter. The major thing is that in the init() call to the
reporter it manually calls functions to initialize the resampling
groups which needs the decision enum and the datatype mapping for the
instructions. THose pieces of informatin are kept in the clone_merge
module and are set in the resampler, which probably will become part
of the API if you want to implement HDF5 pipeline.

These values are then grabbed out of the resampler and fed into the
constructor of the Reporter which saves them as attributes until it
needs them to initialize the HDF5 class.

Also changed "in x.keys" to the more pythonic "in dict" which is
faster and better.