restructured examples, wexplore example runs
No boundary conditions yet.

I restructured the examples folder to avoid duplication of files. I
haven't applied to old examples just the new one I was working
on. Also made a script that generates the minimized file and the extra
files we need for starting the simulation.

fixed some things in the WExplore2 resampler code to accomodate things