added open function, fixed things to run with BCs
Had to include the imports and typos etc for getting the unbindign BCs
to work.

Fixed an error where I had set the closed Flag to False when it should
be True after initialization.

Added a function to open the object in the old way without using the
context manager.