added support for auxiliary data in resampling adn warping in BCs
This is really only tested on resmpling because right now the only
data from the boundary conditions is when it ubinds which is difficult
to do without more simulation so it should be tested.

Otherwise the resampling stuff worked well so I expect the BC warping
stuff to work the same, as most of the code is copy-paste.