refactored warp_walkers
Moved out key things to their own functions.

THis includes the warp function which actually warps a walker and
returns the data and record for it.

Also the check function returns boundary data.

This makes a new aux data model for data that is collected every cycle
when the BC checks the walkers, e.g. the minimum distance, and data
that is collected only when a warp occurs.

This is a good distinction to hard code.