fixed bug where weight of warped walker is set to init_state
The initial `OpenMMWalker` object was simply being copied and used as
the new walker, when really the state was all that was needed. In the
`warp` method of `UnbindingBC` the interface says that it is expecting
a "state" but does no type checking to make sure it is a state,
because that can be whatever. This is why the bug was not caught. In
the main scripts we simply passed in the whole initial walker for the
`initial_state` kwarg. In reality we should only be passing in the
"state" of that walker.

I have modified the sEH-TPPU WExplore2 GPU example to reflect this.