made BCs optional in the sim_manager, major bugfix
If you don't want BCs that wasn't possible now it is. Just don't pass
one in and leave it as None. If it's None it will ignore running that
and just apss along the new_walkers from the segment.

A MAJOR BUGFIX because I just found it doing this was that the
new_walkers were being passed to the resampler and not the
warped_walkers!!!!

I had thought I double checked this but I should have triple-checked
this. This may clear up a lot of issues....