UNITS and UNIT_NAMES, no units handling
The user not giving the reporter any units was not handled. It now
just sets an empty dict which will not set any units in the hdf5.

However we want to set units and I want it to be easy for openmm so I
had a constant in that module that gave the simtk.units units for each
unit key. Howver, this meant you had to manually get the names of the
units for serialization, which the hdf5 module does not and should not
do (at least until a suitable units library is incorporated and
integrated (married)). So we want to just send in the simple unit
strings.

ALso the UNITS constant had some empty dictionaries I wanted to get
rid of. So I just got rid of those keys. The unit keys are still known
in the hdf5 module and will be treated appropriately if they are given
to the reporter so they don't need to be included in the constant for
initialization which would requrie some token which is nice to avoid.

So I made another constant which is just that in openmm. So just
import that and cast it to a dict and set it in your reporter.

Example is in the Lennard Jones script