openmm walker feature vectors, nowater top, to_mdtraj fix
I noticed that in my simulations the time and 1D values were stored as
1D datasets/arrays.

While this seems okay it makes them a fundamentally different kind of
dataset from say the positions. Position datasets have 3 dimensions, 1
for the frame and 2 for the actual coordinates. This is the form of a
"feature vector" as you would find in other applications. I have seen
this in scikit-learn from which I think it is a good idea to model
after.

While storing a 1D array as a 2D array seems more complicated it makes
sense because it can now be treated the same as all the other fields,
e.g. positions. So you can use the same general methods on both
without having to distinguish between them.

So as in scikit-learn it is up to the user to provide feature vectors
even if they are 1D. THis is easy to accomplish with the resize method
in numpy or by just wrapping them in more arrays
`np.array([[single_value]])`.

So I have updated the OpenMMWalker to reflect this change.

If you don't put feature vectors in, HDF5 won't complain you just
won't be able to use the builtin tools. I should make a way to check
but for now you will just have to get it right.

Topology:
I made a nowater json topology for the sEH example I needed for my
analysis but is broadly useful I think.