added new iter_trajs_fields method for new compute_observables
couldn't serialize h5py references so message passing them for
parallel stuff wouldn't work. So I made a new iter function that you
tell it which fields to pass and sends numpy arrays of them back in
the iteration.

This is used in the compute observables so we can do things in parallel