sparse data in TrajHDF5 constructor, _check_compliance_data back
I finished the constructor that is able to handle either flags for
which fields should be sparse and also will take the idxs and figure
out that it should be sparse.

Started on redoing the extend function so that it can handle sparse
data.

This led to the reintroduction of the _compliance function at the
module level, which is now _check_data_compliance because the extend
function uses it to figure out what the new data has.