restructured compute_observables as generator
compute_observables now is a generator which will yield the results
when they are made available. I added the returns results option so
that you can return results and save them.

Furthermore, when save option is turned on it will save results as it
gets them.

Currently the map function must return them in order. ALthough this is
something that I would like to improve on.