sparse traj constructor improvements
testing specifying fields as sparse without giving data for them. The
problem is that the arrays in the HDF5 are not initialized so I need
the rank and shape of the arrays. For some this is standard and others
it is variable to the object.

I fixed a part in the code figuring out what is sparse