initializing fields, defaults, sparse fields with no data
The problem I am trying to solve is initializing the datasets for
sparse fields (with shapes and dtypes known beforehand).

We have a default case for simple MD which I want to make hardcoded
into the module. THis can be lifted out to another layer of
configuration module later but for the foreseeable future this is okay
considering this is customizable to whatever the user wants.

I made a number of widely applicable methods for initializing datasets
for both simple (non-sparse) and sparse fields.

I also advanced the pattern of using the small pathlike field names
for unifying the treatment of simple (not-compound) and compound
fields.

Aside:
Prolly should rename the simple (non-sparse) to contiguous, todo
written. and upcoming

This is accomplished by a method `_get_field_path_grp` which takes in
the pathlike string and returns the group for which the field should
be under (this can be the root trajectory group for simple fields or
compound field group) and the field name it should be created
as (whether its just the simple field name or the compound subfield
name).

This unifies in the sense that you now just need to perform operations
on the returned group instead of handling them as a special case!

I still need to make sure that there is a runtime initialization of
sparse groups for those we don't know the data shape and type of at
the beginning.

These changes are untested!!! so wait for a later commit