testing/debugging for contiguous pre-initialization
Had to change a number of things that unforeseen consequences.

Forgot to add FEATURE_DTYPES which was easy enough as a module
constant.

removed the part where I check to see if the number of dimensions was
passed in and set that vcariable. It is not used and would be deleted
immediately. Furthermore the positions dat is always passed in and
would determine that anyhow making it unnecessary. So I reverted that
to a property method to just look at the positions shape.

Also redid the property method for the other positions shape derived
values n_atoms and n_frames. They also had their own things that
needed them in the create constructor so I make some temporary hidden
attributes for these that can be used by other methods called from the
create constructor.

The main one being a method that initializes the feature shapes and
feature dtypes to the defaults if they weren't given.

Fixed a bug where the sparse_field_paths were not being saved in the
right list and then another list of sparse_field_paths were being used
to set the flags.

Moved the setting of the closed attribute to False before the first
context managed section so I stopped getting the error when it crashes
that no attribute exists since I was setting it after the first
context.

Moved all checking of input traj_data to the constructor to the
create_init constructor where it should be. This goes with the moving
of the setting of system specific attributes as well.

Renamed _init_simple_field to _init_contiguous_field so there isn't
two types of 'simple' fields (non-compound and non-sparse).

Added a special case in the _extend_field method (called in  extend)
to handle when the array was just initialized from all this new stuff
and is actually empty. First resizes to the incoming number of frames
and moves the feature shape dimensions up to their maxshape then sets
the data there.

Not done for the other methods because some reworking is to be done to
utilize the pathlikes pattern.