initialization and create_init for WepyHDF5 sparse trajs
copied where I could from TrajHDF5. Some differences are that I had to
get the number of atoms from the topology so there is a new function
`_json_top_atom_count` to deal with counting the atoms in the json
string.

I also had to hardcode the number of dimensions in the set default
function.

The major difference is that I need to be able to store the shapes,
dtypes, and sparse flags in the HDF5 in case they are never written
the first time the object is instantiated for the dataset. So those
are created in the new top level group '_settings'.