remove broadcasting of func args, compute_obs not generator
I removed automatic broadcasting of args to mapped funcs in
traj_fields_map. Dropped that because that is actually a problem. I
wasn't handling generators properly.

Also the compute_observables function had an optional yield statement
buyt that was turning the whole thing into a generator so you had to
iterate to even make it have side-effects... So it just collects a
list and less efficient. But I guess that is why we save to the file
in the first place!