finished redoing functions TrajHDF5 adn WepyHDF5
I believe my work here is almost done!!

Refactoring etc. everything to the new way made transferring
strategies from TrajHDF5 to WepyHDF5 very easy and a lot is seriously
just copy-pasted. That means that moving that functionality into it's
own module will be very easy!!

So the two classes are newarly the same in certain functions and it
seems that sparse_trajs is working with WepyDHF5 after a few tests
with just extending trajectories and passing in sparse idxs to
add_traj.

next step will be to do more thorough testing then to make a reporter
that does sparse trajectories and test in actual simulations