sparse trajectories created from the reporter
just added 3 options for the sparse fields.

2 are just the shapes and dtypes that will get passed to the
constructor.

The `sparse_fields` kwarg is a dictionary of the sparse field and the
cycle saving frequency.

I also added a `save_fields` kwarg to specify the fields you want to
save at all. Previously it would just save all of them that the walker
had. so now you can exclude stuff.