multiple reporters, walkers pickle reporter
I added the ability of the sim_manager to have multiple reporters so
that you can report in a ny number of ways without having to write new
reporters that combine functionality.

I also made the reporter that Alex wanted for saving pickles by
copying his code on his branch. This was useful for testing too.

I added this to the Lennard Jones example.