first draft of sparse atoms
I implemented in the reporter a way to easily set that you want to
save the whole system.

Frequencies of saving are not handled yet.

I pass the alt_reps and the main representation slice to the
constructor of WepyHDF5 and initialize the HDF5 groups/datasets for
saving these.

In add_traj these are handled and positions are sliced before adding
to the trajectory fields.

NOt tested.