refactored project to deal with new resampling module organization
Finished implementing the ABCs for:
- novelty
- decision
- decider
- resampler

Now there are 5 well defined submodules of resampling (which I will
call the resampling framework or something similar):
- distances
- novelty
- decisions
- deciders
- resamplers

The architecture of this is so that you can create resamplers just by
composing novelty assigners (from `novelty`) and deciders (from
`deciders`) without even having to subclass `Resampler`.

Of course you can and there is an example where it automatically
composes two examples of a novelty assigner and a decider.

Now the `distances` and `decisions` modules can be used in writing
your own resampler that does not use the novelty->decider framework,
which is useful for prototyping something that may not fit cleanly
into that abstraction (perhaps WExplore2).

However, if the Novelty->Decider framework is used the distances and
decisions modules are supports to those respectively.

For instance your NoveltyAssigner may be a bin
assigner (e.g. WExplore) that needs a distance function. That way You
can have the same bin assigner and just plug in different distance
functions.

For the Decider you may have some sort of balancing algorithm (e.g.
WExplore) that changes how decisions are made but the same types of
decisions are being made, which operate in the same way (i.e. cloning
and merging). So that you can just use the core `CloneMergeDecision`
class in `decisions` which implements the enumeration of the
decisions, mappings from their values to string names, the record
types (namedtuples) for instruction records as well as their numpy
dtype equivalents, a method for easily generating
ResamplingRecords (a list of which are returned to the simulation
manager and reporter for building lineages), and importantly an
`action` method which given waklers and a list of resampling records
will return walkers with the given decisions and instructions applied.

This final method was a surprise to be able to put here but it makes
sense as you are tying together the decision types, their
instructions, and the method of acting on those instructions (and
specification really) in one place. They are all logically tied
together and that means a decider (or more monolithic resampler) can
focus on just generating the decisions and not the code to do the
clones and merges.

Additionally this architecture makes it possible to override this
action to some other behavior while inheriting all the other useful
stuff. (say you want to choose how to keep state in a merge
differently).

There is a WIP for a RandomCloneMergeDecider underway but it isn't
that important and I was overcomplicating it so I will wait to make
that, but the concept should be clear.

I also modified the wexplore2 resampler slightly to incorporate these
changes, particularly this effects the calls to the decision
enumeration and the generation of the resampling records, which is
easier I think. Because the Decision can be set as a class variable
and its name shortened.

One change in WepyHDF5 which deals with the type of the
`instruction_dtypes_tokens` in the init_run_resampling method. This
will now accept the raw enumeration types and deal with them how it
needs to internally. This changes the way they are passed in so that
now you just give it the whole enum.

Moved openmm.py to a new wepy.runners.openmm module.

Other changes:

Removed the protoypes folder from gitignore so I can add those if I
want to.

Moved the parent table and parent panel stuff out of the old
wepy.resampling.clone_merge module into the prototypes and committed,
that way they are in the main branch.

Then I deleted that old clone_merge module as the components are in
the component libaries elsewhere.

Added prototypes for testing the decider etc. they are small and
people may want to see them.

Split the Lennard Jones example into the random clone merge example
and the wexplore example.

Removed some druft from the sim_manager