modified the interface for Instruction constants
To account for the variable length instructions I restructured the
class constants a bit to allow for uniformity and removed duplication
in the way that the fields were defined.

Now though to get what previously was INSTRUCTION_RECORDS you use a
method instruction_records() and for INSTRUCTION_DTYPES you use
instruction_dtypes()

And it will automatically return dictionaries as well.

Moved the VariableLengthRecord to the Decision base module which is
where the record is implemeneted anyhow