Fixed resampler code
Given the previous work of Alex on pythonifying his code and making it
work within the wepy framework of recording decisions and isntructions
I set to fix the bugs from that and to incorporate it into the
framework even more.

I started with trying to get the Lennard Jones example working. I
found my walkers weren't any different from each other so I changed
teh integrator to a Langevin integrator and made the seed forcibly set
to 0 which is a special value that chooses another random number
internally. This seemed to fix that problem.

I ended up renaming the novelty stuff to `Scorer` so you call
scorer.score. THis is much to the behest of other parties but I don't
see a better way to name this that won't fill up the module with too
many odds and end type things.

The Decider is not implemented outside of the Wexplore2 resampler and
is just internal for now.

But the Scorer was needed because in addition to wanting to customize
the distance metric, we want to be able to customize how the all to
all distance calculations are done.

For instance in my Lennard Jones example I simply just need to apply
the distance metric in an all to all mapping. Thats it. So there is a
base `AllToAllScorer` that has a hidden method (so that it can be
inherited without being overriden in the `score` function) to do this
when `score` is called by the resampler.

But for molecular systems it is common to want to use the rmsd
distance function between conformations but to do this alignment needs
to be performed first. It is very inneficient to do so all inside the
distance function so we want to align them all first to a single
reference (N) instead of doing it for each pair in the all to all
mapping (N^2). So we simply would only need to subclass the
`AllToAllScorer` and add a line for superposing all the walkers before
calling the hidden all_to_all method it inherited.

That is the way it is at least.

There were some fixes and debugging that had to be done to the
resampler and the associated functions. To help me do this I
refactored some variables to make their meaning more clear (to me at
least).

I also had to fix some stuff related to making records:
- implement __len__ in VariableLengthRecord
- bug checking for Ellipsis

Now all that needs to be done is make the fixes in the hdf5 module
that will allow for saving the variable length records valid as it
seems to be working.