finished resampling panel method in hdf5
This method does the task of converting the flat table of records from
the HDF5 of the form (cycle_idx, step_idx, walker_idx, decision_id,
instruction) to a hierarchical form of a panel, which is a collection
of tables that are the same size. Named after the multidimensional
dataframes in pandas called panels. Maybe the name isn't that
intuitive but it works for me and could be renamed later.

The new format is a list of cycle collections for each cycle in the
run. In each cycle collection is another list of steps. A step is a
container that contains (in the order of the walker_idxs) the
resampling records as defined in the decision module,
i.e. (decision_id, instruction).

If a walker is missing from a step the record will be None.

This makes the task of making parent tables etc. much more
straightforward as they don't need to parse the flat data structure
and simple nested looping routines are easy to perform.