Moved parent panel and table to clone merge decision
THis is because these functions work specifically on the details of
this Decision class:

- what does clone and merge mean?
- which decisions leave children behind?
- which values in an instruction are the children walkers?

I suppose the net_parent table doesn't because it takes an already
formed parent panel, but I suppose we can refactor that somewhere else
later.

IN any case they work, for at least single step parent panels.

The main difference is that I have left out the row for the initial
walkers so the parent tables have the same number of rows as the
number of cycles. This does mean that the final "walkers" in the tree
exist but are the result of running a segment followed by resampling
as opposed to simply the results of the segment. So it is uniform, but
is kind of unnecessary perhaps.