get traj fields update, trace methods
In order to implement the get_trace_fields method (and the mdtraj
trace method) I upgraded the _get_traj_field methods for both
contiguous and sparse to allow for getting a selection of frames.

THis allowed for a simple but maybe slow implementation of the
get_trace_fields that allows you to choose a number of fields and a
list of (cycle_idx, walker_idx) tuples that can be used to select walkers.