example to test "scorer" assignments, rework assignments
Reworked the assignments algorithm and broke into pieces and removed
recursivity.

Recursivity is not really necessary as you are always just finding a
path to the bottom of the tree and you never need to come back up a
level. Even when you are creating new regions.

We made the new region creation into more of a banching process as
well.

That way you can cleanly break the pure assignment problem (a walker state can
always be assigned to a region) and the optimization problem of
creating a useful number of regions.

So the main subroutine of the "scorer" is place_walker which itself
contains a subroutine to strictly assign a walker and then test the
distances to ascertain whether or not making new regions is necessary.

We also reworked the way that the images are stored. Previously a node
contained the actual image itself. This was redundant because all the
non-leaf nodes have images which are contained in the images of the
leaf nodes.

So we made a separate list of the actual images so they are only
stored once, and then only the image index needs to be stored in the
node.

A further optimization that could be run is to cache distance
calculations for specific images, so that during the n-ary search to
find the leaf node region the walker is in it need not repeat the
distance calculation which would be present at all levels.

THis was not done but should be.

This was tested with a simple model for 1 step and seems to work properly.

OTher changes:

Example of using Nazanin's profiler (really a perfomance tester) with
an example.

Prototype for testing the components of the WExplore1Resampler.

Added RunCycleSlice to main hdf5 module.