new abstract Walker and WalkerState classes and API
THis marks the introduction of the API for WalkerState. That is it is
supposed to have attributes like a dictionary.

SO that any interface can just access them by keys instead of having
to know the specific calls to the methods and properties.

This also breaks a lot of behavior where the Walker was used as a
proxy for the state when really all that was needed was the state.

I fixed this where it was needed in the Lennard Jones example which
included teh Unbinding boundary conditions class mostly.

This means that the runner specific Walkers are a little obsolete as
you just need to have a custom state that is subclassed from State for
it to be valid (or really just __getitem__ implemented :P).

So OpenMMWalker still exists but it just adds an assertion that the
state is an OpenMMState, which is a customized WalkerState. SO I don't
even use it in my example.

Another benefit is that the abstract State class is very useful
because you can just pass the attributes as kwargs and it will handle
it.

ANother part of the API is to pass the whole thing back as a
dictionary. Since WalkerState does not inherit from abstract
dictionary, we just make this a method called dict().

THis makes sense because potentially this could require some
processing, as in the case of converting the simtk unit arrays to
numpy arrays.

There are still broken parts of RandomWalk but should be trivial to
solve for Nazanin.