bug in HDF5 reporter that modified walkers
In the WepyHDF5Reporter the data from the walker was modified without
making a copy which led to errors in the randomwalk example where the
number of axes in the positions was increased. THis wasn't apparent
for the OpenMMState because it by default of accessing the attributes
makes copies of the data, because it is converting from simtk.unit
Quantities to numpy arrays.

Was solved by just making a deepcopy of the walker and proceeding with
the same code