fixed a lot of errors in wexplore1
There was a lot of errors surrounding the choice of the walkers to be
cloned and merged. It is too much to go into detail of everything but
it seems to be working stably now.

There is one change which might effect the algorithm that should be
noted. That is the choice of walkers to be cloned.

When a positive balance is handed to a lead for cloning the way it was
supposed to be cloned was... I'm not sure. In any case what I have
implemented is that the clones will be evenly distributed as much as
possible among the cloneable walkers. That is first each cloneable
walker (weight >= pmin) is balance / n_cloneable_walkers. Then the
remainder is given to the highest weighted walkers to be cloned.

THis does NOT prohibit the cloning of walkers into other walkers which
are below pmin!!!!

This does however get pretty close because anything under pmin will
never get cloned again so it won't get any smaller than pmin/2.

Will have to discuss this but pretty sure it is okay.