WorkMapper Object arg, add_observable methods, sliding window function
IMPORTANT API CHANGE: instead of passing in a mapping function you
must pass in the whole workmapper object that has a map method.

Unfortunate to make this API change here but it is happening ;)

In other more dedicated news...

Added methods for adding fields and observables to an HDF5 object.

This included some new helper functions and hidden functions for the
API function `add_observable` and `add_run_abservable`.

- traj_n_frames
- run_n_frames
- _add_run_field :: lets you set any field in a run, with sparse idxs
- _add_field :: set any field for the whole object
Same as the last two but exposed and only for setting in the observables group:
- add_run_observable
- add_observable

Can specify the runs you want (run_idx, traj_idx) tuples for in `run_traj_idx_tuples`

In the tree module added a method for generating sliding window traces
from a parent matrix.