Comments on changes I did on wexplore1
Changes in wexplore1

1)place_walkers():
1-1) The code have not updated the some nodes
"n_mergeable" values.  This was due the wrong number of iteration in
for loop (line 292). I modified the range of loop to (self.n_levels).

1-2) The counting of "n_mergeable" walkers was wrong. It was one less
than correct value so I changed line 288 to consider this
deficit. However, this change did not take into account that the
condition that the number of mergeable walkers should be less than
number of walkers so I add i+1 < len(weights) condition to it.

2)minmax_beneficiaries(): 2-1) Changes in both if conditions:(Not
max_child_idx) do not work in python so I changed it to (max_child_idx
is None).

2-2) The max walker and min walker was selected based on the number of
walkerds that they had ,however, it should have done based on number of
shares that the walkers have.

3)resample_region():

3-1)There was a extra indent in line 390. The checking of
parent_balnce should happen after passing parent balance to the children.

3-2) In the case that parental balance was less than zero, the way we
finding minimum debt a child caould pay was wrong.Since parental
balance in this case was a negative number we needed to take absolute
value of it.

4)balance_tree(): In this function we calling settle_balance function
for each leaf node. I thought it is better to be called for nodes that
their balance is not zero. This is just a performance issue.